it's impossible to have a tab with no selection
tabs default to setting to the first one if not specified, now that you
can specifically ask for one to be set by index, it stands to reason
that you could give this a non-sense input, asking it to select a panel
which doesn’t actually exist.  It seems that in this case it should
never display -no- panel, so I’ll treat that as the same ‘none
specified’ case and show the first one.