add angular directive version
It turns out that managing custom element upgrades in the angular
lifecycle is especially hard for panel sets that are dynamically
created by angular - trying to do so is making the prototype really
screwy.

Adding an implementation of panels as an angular directive with 1:1
match of API surface.  Custom elements and angular directives are
indistinguishable (can be anyway) from the markup perspective and the
main API perspective - as much as I don’t like saying this, it will
actually be easier and cleaner to manage two implementations.